services/trade_producer/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Inigo <inigo_ocariz@me.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
quixstreams = "^3.7.0"
websocket-client = "^1.8.0"
loguru = "^0.7.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
extend-select = ["I"]